backend/dfe_api/requirements.txt
# API DFE + dependencias del conector ARCA
-r ../arca_ve_connector/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-auth>=2.22.0
